requirements.txt
Django             == 4.2.16 
djangorestframework== 3.15.2 
pip                == 20.0.2 
pkg-resources      == 0.0.0  
setuptools         == 44.0.0 
sqlparse           == 0.5.1  
typing-extensions  == 4.12.2 